Multiple changes, see description  - Created teleop, autonomous, and misc packages within commands package  - Moved all of the commands to their respective packages  - Removed unnecessary comment in ElevatorSubsystem class
