+++ src/main/java/frc/robot/Robot.java
@@ -6,10 +6,10 @@ import edu.wpi.first.wpilibj.RobotController;
 import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj.command.Scheduler;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import frc.robot.commands.CompressorCommand;
-import frc.robot.commands.ElevatorCommand;
-import frc.robot.commands.IntakeCommand;
-import frc.robot.commands.TeleopCommand;
+import frc.robot.commands.misc.CompressorCommand;
+import frc.robot.commands.teleop.ElevatorCommand;
+import frc.robot.commands.teleop.IntakeCommand;
+import frc.robot.commands.teleop.TeleopCommand;
 import frc.robot.helpers.Constants;
 import frc.robot.subsystems.DriveSubsystem;
 import frc.robot.subsystems.PneumaticSubsystem;
