+++ src/main/java/frc/robot/commands/teleop/FineTurningCommand.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.teleop;
 
 import edu.wpi.first.wpilibj.command.Command;
 import frc.robot.helpers.Constants;
