+++ src/main/java/frc/robot/commands/teleop/VisionCommand.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.teleop;
 
 import static frc.robot.OperatorInterface.mainStick;
