+++ src/main/java/frc/robot/subsystems/ElevatorSubsystem.java
@@ -1,9 +1,8 @@
 package frc.robot.subsystems;
 
 import com.ctre.phoenix.motorcontrol.can.WPI_TalonSRX;
-// import com.ctre.phoenix.motorcontrol.can.WPI_VictorSPX;
 import edu.wpi.first.wpilibj.command.Subsystem;
-import frc.robot.commands.ElevatorCommand;
+import frc.robot.commands.teleop.ElevatorCommand;
 import frc.robot.helpers.Map;
 
 public class ElevatorSubsystem extends Subsystem {
